free_agent/kdm_hotel/requirements.txt

Flask
pymongo
langchain
chromadb
pypdf
python-dotenv
ollama
